requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
Naked==0.1.31
pick==1.2.0
pycodestyle==2.8.0
pycryptodome==3.14.1
pycryptodomex==3.14.1
pyotp==2.6.0
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
toml==0.10.2
urllib3==1.26.9
